Create folder components and folder types.
