+++ my-app/src/App.tsx
@@ -1,4 +1,6 @@
 import "./App.css";
+import UserList from "./components/UserList/UserList";
+import User from "./types/User";
 
 const USERS: User[] = [
   { id: "1", name: "thanh", position: "$1", startDate: "12 Nov 2024" },
@@ -6,45 +8,6 @@ const USERS: User[] = [
   { id: "3", name: "nguyen", position: "$1", startDate: "12 Nov 2024" },
   { id: "4", name: "anh", position: "$1", startDate: "12 Nov 2024" },
 ];
-
-interface User {
-  id: string;
-  name: string;
-  position: string;
-  startDate: string;
-}
-
-interface UserRowProps {
-  user: User;
-}
-
-interface UserListProps {
-  users: User[];
-}
-
-const UserRow = ({ user }: UserRowProps) => {
-  const { name, position, startDate } = user;
-
-  return (
-    <tr>
-      <td>{name}</td>
-      <td>{position}</td>
-      <td>{startDate}</td>
-    </tr>
-  );
-};
-
-const UserList = ({ users }: UserListProps) => {
-  return (
-    <table>
-      <tbody>
-        {users.map((user: User) => (
-          <UserRow key={user.id} user={user} />
-        ))}
-      </tbody>
-    </table>
-  );
-};
 
 const App = () => {
   return <UserList users={USERS} />;
