+++ my-app/src/components/UserRow/UserRow.tsx
@@ -1,0 +1,14 @@
+import UserRowProps from "../../types/UserRowProps";
+
+const UserRow = ({ user }: UserRowProps) => {
+  const { name, position, startDate } = user;
+  return (
+    <tr>
+      <td>{name}</td>
+      <td>{position}</td>
+      <td>{startDate}</td>
+    </tr>
+  );
+};
+
+export default UserRow;
