+++ my-app/src/types/User.ts
@@ -1,0 +1,8 @@
+interface User {
+  id: string;
+  name: string;
+  position: string;
+  startDate: string;
+}
+
+export default User;
