+++ my-app/src/types/UserListProps.ts
@@ -1,0 +1,7 @@
+import User from "./User";
+
+interface UserListProps {
+  users: User[];
+}
+
+export default UserListProps;
