+++ my-app/src/types/UserRowProps.ts
@@ -1,0 +1,7 @@
+import User from "./User";
+
+interface UserRowProps {
+  user: User;
+}
+
+export default UserRowProps;
